refactor: split victory overlay, political borders, and map prompt (#3878)
Extract victory panel corner brackets, province vs faction political border
paint, and the work-target selection prompt into dedicated modules for Phase 3
flame map modularization without changing public behavior.
