+++ SPEC/program/repo-and-packages.md
@@ -22,10 +22,10 @@ Monorepo layout:
 - **`features/game/flame/caches/`** — shared Flame asset decode helper (`AssetImageCache`) and map marker icon caches (town, port, fleet, civilian, resource, province label). Public barrel: `caches/caches.dart`.
 - **`features/game/flame/map_area/`** — `GameMapAreaBackground` and other render-only map chrome. Public barrel: `map_area/map_area.dart`.
 - **`features/game/flame/map_state/`** — `GameMapArea` widget stack, selection, province actions, turn-feed hooks, and draft projections. `game_map_area.dart` library parts: `game_map_area_turn_feed_labels.dart` (turn-feed display labels + map-locate helpers), `game_map_area_turn_feed.dart` (turn-feed entry builder), `game_map_area_build_map_stack.dart` (map canvas stack + in-map overlay chrome), `game_map_area_build_overlays.dart` (play-area shell + debug/narrow slots), `game_map_area_build.dart` (controls bar + play-area shell). `game_map_area_state_logic.dart` library parts: `game_map_area_state_logic_work_targets.dart`, `game_map_area_state_logic_draft_projection.dart`, `game_map_area_state_logic_province_actions.dart`. `game_map_area_province_action_states.dart` library parts: `game_map_area_province_action_states_prospect.dart`, `game_map_area_province_action_states_explore.dart`, `game_map_area_province_action_states_build_improvement.dart`. `game_map_area_fleet_draft_projection.dart` library parts: `game_map_area_fleet_draft_projection_scope.dart` (location scope keys), `game_map_area_fleet_draft_projection_tiles.dart` (drawable tile resolution). Public barrel: `map_state/map_state.dart`.
-- **`features/game/flame/region_map/`** — `RegionMapComponent` Flame render stack and viewport helpers. `CtRegionMapGame` host (`ct_region_map_game.dart`) library parts: `ct_region_map_game_camera.dart` (clamp, zoom sync, resize), `ct_region_map_game_props.dart` (live prop sync), `ct_region_map_game_viewport.dart` (public pan/zoom/tap/hover API). `region_map_component.dart` library parts split shared constants/plate layout (`region_map_component_shared.dart`) from visibility modes (`region_map_component_shared_visibility.dart`), fog/transport helpers (`region_map_component_shared_visibility_fog_transport.dart`), extraction disc helpers (`region_map_component_shared_visibility_extraction.dart`), and label/halos (`region_map_component_shared_visibility_*.dart`); hover/tap hit-testing (`region_map_component_interaction.dart`); civilian vs fleet unit marker paint (`region_map_component_render_markers_units_civilian.dart`, `region_map_component_render_markers_units_fleet.dart`); capital rings (`region_map_component_render_markers_settlements_capitals.dart`), town/port paint (`region_map_component_render_markers_settlements_towns.dart`), and warp-zone edges (`region_map_component_render_markers_settlements_warp.dart`); province label compute vs paint (`region_map_component_render_political_labels_province_compute.dart`, `region_map_component_render_political_labels_province_paint.dart`). Public barrel: `region_map/region_map.dart`.
+- **`features/game/flame/region_map/`** — `RegionMapComponent` Flame render stack and viewport helpers. `CtRegionMapGame` host (`ct_region_map_game.dart`) library parts: `ct_region_map_game_camera.dart` (clamp, zoom sync, resize), `ct_region_map_game_props.dart` (live prop sync), `ct_region_map_game_viewport.dart` (public pan/zoom/tap/hover API). `region_map_component.dart` library parts split shared constants/plate layout (`region_map_component_shared.dart`) from visibility modes (`region_map_component_shared_visibility.dart`), fog/transport helpers (`region_map_component_shared_visibility_fog_transport.dart`), extraction disc helpers (`region_map_component_shared_visibility_extraction.dart`), and label/halos (`region_map_component_shared_visibility_*.dart`); hover/tap hit-testing (`region_map_component_interaction.dart`); province topology borders and GP ownership tint (`region_map_component_render_political_borders_province.dart`); faction political borders (`region_map_component_render_political_borders_faction.dart`); civilian vs fleet unit marker paint (`region_map_component_render_markers_units_civilian.dart`, `region_map_component_render_markers_units_fleet.dart`); capital rings (`region_map_component_render_markers_settlements_capitals.dart`), town/port paint (`region_map_component_render_markers_settlements_towns.dart`), and warp-zone edges (`region_map_component_render_markers_settlements_warp.dart`); province label compute vs paint (`region_map_component_render_political_labels_province_compute.dart`, `region_map_component_render_political_labels_province_paint.dart`). Public barrel: `region_map/region_map.dart`.
 - **`features/game/flame/tilesets/`** — terrain Wang tilesets, transport overlay tilesets, and connectivity masks. `terrain_tileset.dart` library parts: `terrain_tileset_models.dart` (Wang tile types, layer enum), `terrain_tileset_variant_keys.dart` (L2 variant tile-key resolution). Public barrel: `tilesets/tilesets.dart`.
 - **`features/game/flame/controls/`** — in-game map shell chrome: top/tab bar ([GameMapControls](../../app/lib/features/game/flame/controls/game_map_controls.dart)), corner tool row ([GameMapCornerControls](../../app/lib/features/game/flame/controls/game_map_corner_controls.dart)), and empire left rail ([GameMapEmpireLeftRail](../../app/lib/features/game/flame/controls/game_map_empire_left_rail.dart)). Public barrel: `controls/controls.dart`.
-- **`features/game/flame/overlays/`** — map and session overlays/dialogs: province detail hosts ([GameMapProvinceDetailSidePanel](../../app/lib/features/game/flame/overlays/game_map_province_detail_side_panel.dart), [GameMapNarrowDetailOverlaySlot](../../app/lib/features/game/flame/overlays/game_map_narrow_detail_overlay.dart)), turn confirmation/processing dialogs, victory overlay, exit confirm, and debug console panel. Public barrel: `overlays/overlays.dart`.
+- **`features/game/flame/overlays/`** — map and session overlays/dialogs: province detail hosts ([GameMapProvinceDetailSidePanel](../../app/lib/features/game/flame/overlays/game_map_province_detail_side_panel.dart), [GameMapNarrowDetailOverlaySlot](../../app/lib/features/game/flame/overlays/game_map_narrow_detail_overlay.dart)), turn confirmation/processing dialogs, victory overlay ([VictoryPanel](../../app/lib/features/game/flame/overlays/victory_overlay.dart) library parts: `victory_overlay_panel.dart`, `victory_overlay_panel_corners.dart`), exit confirm, and debug console panel. Public barrel: `overlays/overlays.dart`.
 - **`features/game/flame/minimap/`** — [GameRegionMinimap](../../app/lib/features/game/flame/minimap/game_region_minimap.dart) widget and viewport math helpers; `part` files: `game_region_minimap_controls.dart` (zoom slider + toggle chrome), `game_region_minimap_painter.dart` (terrain fill + viewport indicator). Public barrel: `minimap/minimap.dart`.
 - **`features/game/flame/render/`** — pure map render helpers: GP ownership tint layer, warp-zone edge geometry, and resource extraction disc palette table. Public barrel: `render/render.dart`. `map_area/` and `region_map/` must not import `map_state/`; `map_state/` may import `map_area` public exports only (`map_area/map_area.dart`). Legacy re-export shims under `flame/` root are forbidden — import submodule barrels directly.
 - **Widgetbook catalogs** (`widgetbook_host/lib/catalogs/`) — dev-only Widgetbook stories and catalog `part` files. `app/lib/widgetbook.dart` is a thin re-export shim for local `flutter run -t lib/widgetbook.dart`; production analyze scope for `app/lib/features`, `app/lib/core`, and `app/lib/widgets` excludes relocated catalogs (`repo.app_widgetbook_catalog_location`).
