+++ app/lib/features/game/flame/region_map/region_map_component.dart
@@ -40,7 +40,8 @@ part 'region_map_component_render_core_overlays.dart';
 part 'region_map_component_render_political_labels_province_compute.dart';
 part 'region_map_component_render_political_labels_province_paint.dart';
 part 'region_map_component_render_political_labels_sea.dart';
-part 'region_map_component_render_political_borders.dart';
+part 'region_map_component_render_political_borders_province.dart';
+part 'region_map_component_render_political_borders_faction.dart';
 part 'region_map_component_render_markers_selection.dart';
 part 'region_map_component_render_markers_settlements_capitals.dart';
 part 'region_map_component_render_markers_settlements_towns.dart';
